refactor: add load cards hooks
